Fix N-way associative policy.
The N-way associative policy was previously implemented backwards
compared to what is shown on the Wikipedia page for cache placement
policies. To use the terminology from the wiki page, set associative
caches can be thought of as an NxM matrix where N is the number of sets
and M is the number of cache lines per set. Previously, the N-way
associative policy was using M as the number of sets and N as the number
of cache lines per set. As a result, when an N-way associative policy
was created with N=1, it was acting equivalent to the fully associative
policy instead of being equivalent to the direct mapped policy.

https://en.wikipedia.org/wiki/Cache_placement_policies